Ignore errors for a missing CREATE_NEW_FILE
Because I'm sharing the Notes dir with Nextcloud, one machine might have
deleted the CREATE_NEW_FILE file, and the `rm CREATE_NEW_FILE` will fail
on another machine, which causes the scipt to bail.

Just silently handle the case of a missing CREATE_NEW_FILE

TODO: Work out a better solution for this, as creating a new file might
be impossible on one machine if the other machine has deleted it. Maybe
just not delete CREATE_NEW_FILE.